requirements.txt
alembic==1.4.3
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.20
stripe==2.55.0
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
